feat:新增课程和成绩管理全部逻辑，修改用户管理

+++ ruoyi-student/src/main/java/com/ruoyi/student/domain/Course.java
@@ -35,7 +35,7 @@ public class Course implements Serializable {
     /**
      * 课程学分
      */
-    private Integer courseCredits;
+    private Double courseCredits;
 
     /**
      * 课程学时
